Move records blueprint into module and clean imports
