+++ views/records/record_views.py
@@ -1,19 +1,14 @@
 import json
-import csv
-import io
 import logging
-from flask import Blueprint, render_template, abort, request, redirect, url_for, jsonify, current_app, Response
-from db.database import get_connection
-from db.validation import validate_table
+from flask import Blueprint, render_template, abort, request, redirect, url_for, jsonify, current_app
+
 from db.records import (
-    get_all_records,
     get_record_by_id,
     update_field_value,
     create_record,
     delete_record,
     count_nonnull as db_count_nonnull,
     field_distribution,
-    count_records,
     append_edit_log,
     get_edit_history,
     get_edit_entry,
@@ -25,7 +20,6 @@ from db.schema import (
     get_field_schema,
     update_layout as db_update_layout,
     update_field_styling as db_update_field_styling,
-    get_title_field,
 )
 from db.dashboard import sum_field as db_sum_field
 from db.config import get_layout_defaults
@@ -33,238 +27,14 @@ from db.config import get_relationship_visibility, update_relationship_visibilit
 from utils.field_registry import get_field_type, get_type_size_map
 
 records_bp = Blueprint('records', __name__)
+from utils.records_helpers import require_base_table
 
 logger = logging.getLogger(__name__)
 
 
-def _parse_list_params(table):
-    """Parse query parameters for list/export routes."""
-    base_tables = current_app.config['BASE_TABLES']
-    if table not in base_tables:
-        abort(404)
-    fields = list(get_field_schema()[table].keys())
-    search = request.args.get('search', '').strip()
-
-    raw_args = request.args.to_dict(flat=False)
-    field_map = {f.lower(): f for f in fields}
-    normalized = {field_map.get(k.lower(), k): v for k, v in raw_args.items()}
-
-    filters = {
-        k: v
-        for k, v in normalized.items()
-        if k in fields
-        or (k.endswith('_min') and k[:-4] in fields)
-        or (k.endswith('_max') and k[:-4] in fields)
-        or (k.endswith('_start') and k[:-6] in fields)
-        or (k.endswith('_end') and k[:-4] in fields)
-    }
-
-    ops = {
-        k[:-3]: (v[0] if isinstance(v, list) else v)
-        for k, v in normalized.items()
-        if k.endswith('_op') and k[:-3] in fields
-    }
-
-    modes = {
-        k[:-5]: (v[0] if isinstance(v, list) else v)
-        for k, v in normalized.items()
-        if k.endswith('_mode') and k[:-5] in fields
-    }
-
-    sort_field = request.args.get('sort')
-    direction = request.args.get('dir', 'asc')
-
-    return {
-        'fields': fields,
-        'search': search,
-        'filters': filters,
-        'ops': ops,
-        'modes': modes,
-        'sort_field': sort_field,
-        'direction': direction,
-    }
-
-def _build_list_context(table):
-    params = _parse_list_params(table)
-    fields = params['fields']
-    search = params['search']
-    filters = params['filters']
-    ops = params['ops']
-    modes = params['modes']
-    sort_field = params['sort_field']
-    direction = params['direction']
-    page = int(request.args.get('page', 1))
-    per_page = 500
-    offset = (page - 1) * per_page
-    records = get_all_records(
-        table,
-        search=search,
-        filters=filters,
-        ops=ops,
-        modes=modes,
-        sort_field=sort_field,
-        direction=direction,
-        limit=per_page,
-        offset=offset,
-    )
-    total_count = count_records(table, search=search, filters=filters, ops=ops, modes=modes)
-    args_without_page = request.args.to_dict(flat=False)
-    args_without_page.pop('page', None)
-    from urllib.parse import urlencode
-    base_qs = urlencode(args_without_page, doseq=True)
-    args_no_sort = dict(args_without_page)
-    args_no_sort.pop('sort', None)
-    args_no_sort.pop('dir', None)
-    base_qs_no_sort = urlencode(args_no_sort, doseq=True)
-    total_pages = (total_count + per_page - 1) // per_page
-    start = offset + 1 if total_count else 0
-    end = min(offset + len(records), total_count)
-
-    return {
-        'table': table,
-        'fields': fields,
-        'records': records,
-        'sort_field': sort_field,
-        'direction': direction,
-        'page': page,
-        'total_pages': total_pages,
-        'total_count': total_count,
-        'start': start,
-        'end': end,
-        'per_page': per_page,
-        'base_qs': base_qs,
-        'base_qs_no_sort': base_qs_no_sort,
-        'has_filters': bool(filters),
-    }
-
-@records_bp.route('/<table>')
-def list_view(table):
-    ctx = _build_list_context(table)
-    if request.accept_mimetypes.best == 'application/json':
-        rows = render_template('_record_rows.html', **ctx)
-        pager = render_template('_pagination.html', **ctx)
-        count = render_template('_record_count.html', **ctx)
-        filters = render_template('_filter_chips.html', **ctx)
-        ctx.update({'rows_html': rows, 'pagination_html': pager,
-                    'count_html': count, 'filters_html': filters})
-        return jsonify(ctx)
-    return render_template('list_view.html', request=request, **ctx)
-
-
-@records_bp.route('/api/<table>/list')
-def api_list(table):
-    """Return basic id/label info for records in the table.
-
-    Optional query parameters:
-    - ``search``: substring match on the label field.
-    - ``limit``: maximum number of rows to return.
-    """
-    try:
-        validate_table(table)
-    except ValueError:
-        abort(404)
-
-    search = request.args.get("search")
-    limit = request.args.get("limit", type=int)
-
-    with get_connection() as conn:
-        cur = conn.cursor()
-        cur.execute(f"PRAGMA table_info({table})")
-        cols = [r[1] for r in cur.fetchall()]
-        if not cols:
-            return jsonify([])
-        title_field = get_title_field(table)
-        if title_field:
-            label_field = title_field
-        elif table in cols:
-            label_field = table
-        elif len(cols) > 1:
-            label_field = cols[1]
-        else:
-            label_field = cols[0]
-
-        sql = f"SELECT id, {label_field} FROM {table}"
-        params = []
-        if search is not None:
-            sql += f" WHERE {label_field} LIKE ?"
-            params.append(f"%{search}%")
-        if limit is not None:
-            sql += " LIMIT ?"
-            params.append(limit)
-        cur.execute(sql, params)
-        rows = cur.fetchall()
-
-    return jsonify([{'id': r[0], 'label': r[1]} for r in rows])
-
-
-@records_bp.route('/api/<table>/records')
-def api_records(table):
-    ctx = _build_list_context(table)
-    rows = render_template('_record_rows.html', **ctx)
-    pager = render_template('_pagination.html', **ctx)
-    count = render_template('_record_count.html', **ctx)
-    filters = render_template('_filter_chips.html', **ctx)
-    ctx.update({'rows_html': rows, 'pagination_html': pager,
-                'count_html': count, 'filters_html': filters})
-    return jsonify(ctx)
-
-
-@records_bp.route('/<table>/export')
-def export_csv(table):
-    """Stream CSV of records using current filters and search."""
-    params = _parse_list_params(table)
-    fields = [f for f in params['fields'] if not f.startswith('_')]
-    ids_param = request.args.getlist('ids') or request.args.get('ids', '')
-    if isinstance(ids_param, str):
-        ids = [i for i in ids_param.split(',') if i]
-    else:
-        ids = ids_param
-    ids = [int(i) for i in ids if str(i).isdigit()]
-    if ids:
-        records = get_all_records(
-            table,
-            filters={'id': ids},
-            ops={'id': 'equals'},
-            modes={'id': 'any'},
-        )
-    else:
-        records = get_all_records(
-            table,
-            search=params['search'],
-            filters=params['filters'],
-            ops=params['ops'],
-            modes=params['modes'],
-            sort_field=params['sort_field'],
-            direction=params['direction'],
-        )
-
-    def generate():
-        buf = io.StringIO()
-        writer = csv.writer(buf)
-        writer.writerow(fields)
-        yield buf.getvalue()
-        buf.seek(0)
-        buf.truncate(0)
-        for row in records:
-            writer.writerow([row.get(f, '') for f in fields])
-            yield buf.getvalue()
-            buf.seek(0)
-            buf.truncate(0)
-
-    filename = f"{table}.csv"
-    return Response(
-        generate(),
-        mimetype='text/csv',
-        headers={'Content-Disposition': f'attachment; filename={filename}'}
-    )
-
 @records_bp.route('/<table>/<int:record_id>')
+@require_base_table
 def detail_view(table, record_id):
-    try:
-        validate_table(table)
-    except ValueError:
-        abort(404)
-
     record = get_record_by_id(table, record_id)
     if not record:
         abort(404)
@@ -315,12 +85,11 @@ def detail_view(table, record_id):
         relationship_visibility=visibility_all
     )
 
+
 @records_bp.route('/<table>/<int:record_id>/add-field', methods=['POST'])
+@require_base_table
 def add_field_route(table, record_id):
     try:
-        # Ensure table exists in schema
-        validate_table(table)
-
         field_name = request.form['field_name']
         logger.debug(
             'add_field_route start: table=%r, record_id=%r, form=%r', table, record_id, dict(request.form)
@@ -353,7 +122,9 @@ def add_field_route(table, record_id):
         logger.exception('add_field_route error: %s', e)
         return 'Server error', 500
 
+
 @records_bp.route('/<table>/count-nonnull')
+@require_base_table
 def count_nonnull(table):
     field = request.args.get('field')
     try:
@@ -362,7 +133,9 @@ def count_nonnull(table):
         return jsonify({'count': 0}), 400
     return jsonify({'count': count})
 
+
 @records_bp.route('/<table>/sum-field')
+@require_base_table
 def sum_field_route(table):
     field = request.args.get('field')
     try:
@@ -373,6 +146,7 @@ def sum_field_route(table):
 
 
 @records_bp.route('/<table>/field-distribution')
+@require_base_table
 def field_distribution_route(table):
     field = request.args.get('field')
     try:
@@ -381,7 +155,9 @@ def field_distribution_route(table):
         return jsonify({}), 400
     return jsonify(counts)
 
+
 @records_bp.route('/<table>/<int:record_id>/remove-field', methods=['POST'])
+@require_base_table
 def remove_field_route(table, record_id):
     field_name = request.form.get('field_name')
     fmeta = get_field_schema().get(table, {}).get(field_name)
@@ -391,7 +167,9 @@ def remove_field_route(table, record_id):
     remove_field_from_schema(table, field_name)
     return redirect(url_for('records.detail_view', table=table, record_id=record_id))
 
+
 @records_bp.route('/<table>/<int:record_id>/update', methods=['POST'])
+@require_base_table
 def update_field(table, record_id):
     field = request.form.get('field')
     if not field:
@@ -434,8 +212,6 @@ def update_field(table, record_id):
             str(new_value),
         )
 
-    # For AJAX requests return JSON so the front-end can update without a
-    # redirect. The JS helpers send the "X-Requested-With" header.
     if request.headers.get("X-Requested-With") == "XMLHttpRequest":
         return jsonify({"success": True, "new_value": new_value})
 
@@ -443,6 +219,7 @@ def update_field(table, record_id):
 
 
 @records_bp.route('/<table>/bulk-update', methods=['POST'])
+@require_base_table
 def bulk_update(table):
     data = request.get_json(silent=True)
     if not data:
@@ -481,6 +258,7 @@ def bulk_update(table):
             updated += 1
     return jsonify({'success': True, 'updated': updated})
 
+
 @records_bp.route('/relationship', methods=['POST'])
 def manage_relationship():
     data = request.get_json()
@@ -503,11 +281,10 @@ def manage_relationship():
         logger.info('manage_relationship %s %s:%s %s:%s', action, table_a, id_a, table_b, id_b)
     return {'success': True}
 
+
 @records_bp.route('/<table>/new', methods=['GET', 'POST'])
+@require_base_table
 def create_record_route(table):
-    base_tables = current_app.config['BASE_TABLES']
-    if table not in base_tables:
-        abort(404)
     fields = get_field_schema().get(table, {})
     if request.method == 'POST':
         record_id = create_record(table, request.form)
@@ -517,11 +294,10 @@ def create_record_route(table):
             abort(500, 'Failed to create record')
     return render_template('new_record.html', table=table, fields=fields)
 
+
 @records_bp.route('/<table>/<int:record_id>/delete', methods=['POST'])
+@require_base_table
 def delete_record_route(table, record_id):
-    base_tables = current_app.config['BASE_TABLES']
-    if table not in base_tables:
-        abort(404)
     success = delete_record(table, record_id)
     if not success:
         abort(500, 'Failed to delete record')
@@ -529,6 +305,7 @@ def delete_record_route(table, record_id):
 
 
 @records_bp.route('/<table>/<int:record_id>/undo/<int:edit_id>', methods=['POST'])
+@require_base_table
 def undo_edit_route(table, record_id, edit_id):
     entry = get_edit_entry(edit_id)
     if not entry or entry['table_name'] != table or entry['record_id'] != record_id:
@@ -538,7 +315,10 @@ def undo_edit_route(table, record_id, edit_id):
         abort(500, 'Undo failed')
     return jsonify({'success': True})
 
+
+
 @records_bp.route('/<table>/layout', methods=['POST'])
+@require_base_table
 def update_layout(table):
     data = request.get_json(silent=True)
     if not data or not isinstance(data.get('layout'), list):
@@ -554,6 +334,7 @@ def update_layout(table):
 
 
 @records_bp.route('/<table>/style', methods=['POST'])
+@require_base_table
 def update_style(table):
     """Update styling information for a single field."""
     data = request.get_json(silent=True)
@@ -575,6 +356,7 @@ def update_style(table):
 
 
 @records_bp.route('/<table>/relationships', methods=['POST'])
+@require_base_table
 def update_relationships(table):
     """Update relationship visibility configuration for a table."""
     data = request.get_json(silent=True) or {}
@@ -587,3 +369,6 @@ def update_relationships(table):
         current_app.logger.warning('[relationships] update failed: %s', e)
         return jsonify({'error': 'update failed'}), 500
     return jsonify({'success': True})
+
+# Ensure list view routes are registered after blueprint is defined
+from . import list_views  # noqa: E402,F401
